pyproject.toml
[project]
name = "pizzariafastapi2"
version = "1.0"
description = ""
authors = [
    {name = "Victor Brunner",email = "victor.brunner12@gmail.com"}
]
requires-python = ">=3.10,<4"
dependencies = [
    "alembic (==1.17.2)",
    "annotated-doc (==0.0.4)",
    "annotated-types (==0.7.0)",
    "anyio (==4.12.0)",
    "bcrypt (==4.0.1)",
    "certifi (==2026.1.4)",
    "cffi (==2.0.0)",
    "charset-normalizer (==3.4.4)",
    "click (==8.3.1)",
    "colorama (==0.4.6)",
    "cryptography (==46.0.3)",
    "ecdsa (==0.19.1)",
    "exceptiongroup (==1.3.1)",
    "fastapi (==0.124.4)",
    "greenlet (==3.3.0)",
    "h11 (==0.16.0)",
    "idna (==3.11)",
    "jinja2 (==3.1.6)",
    "mako (==1.3.10)",
    "markupsafe (==3.0.3)",
    "passlib (==1.7.4)",
    "pyasn1 (==0.6.1)",
    "pycparser (==2.23)",
    "pydantic (==2.12.5)",
    "pydantic-core (==2.41.5)",
    "python-dotenv (==1.2.1)",
    "python-jose (==3.5.0)",
    "python-multipart (==0.0.20)",
    "requests (==2.32.5)",
    "rsa (==4.9.1)",
    "six (==1.17.0)",
    "sqlalchemy (==2.0.45)",
    "sqlalchemy-utils (==0.42.1)",
    "starlette (==0.50.0)",
    "tomli (==2.3.0)",
    "typing-inspection (==0.4.2)",
    "typing-extensions (==4.15.0)",
    "urllib3 (==2.6.3)",
    "uvicorn (==0.38.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
